[FLINK-23192][table] Move blackhole connector
