+++ flink-table/flink-table-api-java-bridge/src/main/resources/META-INF/services/org.apache.flink.table.factories.Factory
@@ -14,5 +14,5 @@
 # limitations under the License.
 
 org.apache.flink.connector.datagen.table.DataGenTableSourceFactory
-org.apache.flink.table.factories.BlackHoleTableSinkFactory
+org.apache.flink.connector.blackhole.table.BlackHoleTableSinkFactory
 org.apache.flink.connector.print.table.PrintTableSinkFactory
